Reuse provider entities for more than VPN
- Rename views
- Decouple provider templates and configurations
